double is 64bits, it's float which is 32bits !
svn path=/trunk/koffice/; revision=617019

+++ krita/colorspaces/rgb_float_hdr/kis_rgb_f32_hdr_colorspace.cc
@@ -38,10 +38,10 @@
 KisRgbF32HDRColorSpace::KisRgbF32HDRColorSpace(KoColorSpaceRegistry * parent, KoColorProfile *p)
 : KisRgbFloatHDRColorSpace<RgbF32Traits>("RGBAF32", i18n("RGB (32-bit float/channel) for High Dynamic Range imaging"), parent, RGBAF32FLOAT_LCMS_TYPE)
 {
-    m_channels.push_back(new KoChannelInfo(i18n("Red"), 2 * sizeof(double), KoChannelInfo::COLOR, KoChannelInfo::FLOAT32, sizeof(double), QColor(255,0,0)));
-    m_channels.push_back(new KoChannelInfo(i18n("Green"), 1 * sizeof(double), KoChannelInfo::COLOR, KoChannelInfo::FLOAT32, sizeof(double), QColor(0,255,0)));
+    m_channels.push_back(new KoChannelInfo(i18n("Red"), 2 * sizeof(float), KoChannelInfo::COLOR, KoChannelInfo::FLOAT32, sizeof(float), QColor(255,0,0)));
+    m_channels.push_back(new KoChannelInfo(i18n("Green"), 1 * sizeof(float), KoChannelInfo::COLOR, KoChannelInfo::FLOAT32, sizeof(float), QColor(0,255,0)));
     m_channels.push_back(new KoChannelInfo(i18n("Blue"), 0, KoChannelInfo::COLOR, KoChannelInfo::FLOAT32, sizeof(double), QColor(0,0,255)));
-    m_channels.push_back(new KoChannelInfo(i18n("Alpha"), 3 * sizeof(double), KoChannelInfo::ALPHA, KoChannelInfo::FLOAT32));
+    m_channels.push_back(new KoChannelInfo(i18n("Alpha"), 3 * sizeof(float), KoChannelInfo::ALPHA, KoChannelInfo::FLOAT32));
 
     m_compositeOps.insert( COMPOSITE_OVER, new KoCompositeOpOver<RgbF32Traits>( this ) );
     m_compositeOps.insert( COMPOSITE_ERASE, new KoCompositeOpErase<RgbF32Traits>( this ) );
